docs(exception): add Javadoc comments to TradeValidationException
